fix(merger): Correctly handle keys with non-ASCII characters
When merging keys, characters were read from a `std::string` and implicitly converted to `int` to be used as branch keys in the trie.

On platforms where `char` is signed, characters with a value greater than 127 (e.g., `0x80`) were sign-extended, resulting in a negative integer value. This caused incorrect behavior and merge failures as the trie expects keys in the range 0-255.

This commit fixes the issue by explicitly casting the character to `uint8_t` before its use as a key, ensuring the correct unsigned value is used regardless of the platform's default `char` signedness.

A new test case has been added to verify that keys containing characters with the high bit set are merged correctly.

+++ include/leaves/intern/util/_merger.hpp
@@ -311,7 +311,7 @@ struct _Merger {
     assert(src_split_pos < 256);
 
     if (src.is_leaf()) {
-      int key = split_pos < current_key.size() ? current_key[split_pos]
+      int key = split_pos < current_key.size() ? (uint8_t)current_key[split_pos]
                                                : TrieNode::NONE;
       auto& src_leaf = src.leaf();
       assert(src_leaf->key_size >= src_split_pos);
@@ -341,7 +341,7 @@ struct _Merger {
 
     if (split_pos < current_key.size()) {
       assert(src.is_trie());
-      int key = current_key[split_pos];
+      int key = (uint8_t)current_key[split_pos];
 
       // A new trie with two branches for the old dst and the new src
       // But first, selectively copy the src side — if nothing survives
@@ -662,7 +662,7 @@ struct _Merger {
   typename CursorDst::Transition::trie_ptr expand_trie_with_branch(
       typename CursorDst::Transition::trie_ptr& dst_trie, int suffix_len,
       uint16_t* loffset, std::string& current_key) {
-    int branch_key = suffix_len ? current_key[dst_cursor.current_key.size()]
+    int branch_key = suffix_len ? (uint8_t)current_key[dst_cursor.current_key.size()]
                                 : TrieNode::NONE;
     assert(!(branch_key == TrieNode::NONE ? dst_trie->has_none()
                                           : dst_trie->isset(branch_key)));
